# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediaVault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: getPlaylists
      tags: [playlists]
      summary: List all playlists
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists/{id}/items:
    post:
      operationId: addToPlaylist
      tags: [playlists]
      summary: Add media items to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistItemsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
    put:
      operationId: reorderPlaylist
      tags: [playlists]
      summary: Reorder playlist items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistReorderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistDetail"
  /playlists/{id}/items/{itemId}:
    delete:
      operationId: removeFromPlaylist
      tags: [playlists]
      summary: Remove an item from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Playback / Player ─────────────────────────────────────────────────────
...
  /searches:
    get:
      operationId: getSavedSearches
      tags: [searches]
      summary: List saved searches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedSearch"
    post:
      operationId: createSavedSearch
      tags: [searches]
      summary: Save a search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
  /searches/{id}:
    delete:
      operationId: deleteSavedSearch
      tags: [searches]
      summary: Delete a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Settings ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MediaFile:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [video, audio]
        size:
          type: integer
        duration:
          type: ["number", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        fps:
          type: ["number", "null"]
        bitrate:
          type: ["integer", "null"]
        codec:
          type: ["string", "null"]
        format:
          type: string
        thumbnail:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        favorite:
          type: boolean
        playCount:
          type: integer
        lastPlayed:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, offline]
        dateAdded:
          type: string
        dateModified:
          type: string
        folderId:
          type: ["integer", "null"]
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/MediaMetadataEntry"
      required: [id, path, filename, type, size, format, favorite, playCount, status, dateAdded, dateModified]
...
    MediaMetadataEntry:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        valueId:
          type: integer
        value:
          type: string
      required: [categoryId, categoryName, categorySlug, valueId, value]
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isSmart:
          type: boolean
        smartFilters:
          type: ["string", "null"]
        itemCount:
          type: integer
        totalDuration:
          type: ["number", "null"]
        thumbnail:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, isSmart, itemCount, createdAt, updatedAt]
    PlaylistInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isSmart:
          type: boolean
        smartFilters:
          type: string
        mediaIds:
          type: array
          items:
            type: integer
      required: [name]
    PlaylistUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        smartFilters:
          type: ["string", "null"]
    PlaylistDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isSmart:
          type: boolean
        smartFilters:
          type: ["string", "null"]
        itemCount:
          type: integer
        totalDuration:
          type: ["number", "null"]
        thumbnail:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlaylistItem"
      required: [id, name, isSmart, itemCount, createdAt, updatedAt, items]
    PlaylistItem:
      type: object
      properties:
        id:
          type: integer
        position:
          type: integer
        media:
          $ref: "#/components/schemas/MediaFile"
      required: [id, position, media]
    PlaylistItemsInput:
      type: object
      properties:
        mediaIds:
          type: array
          items:
            type: integer
      required: [mediaIds]
    PlaylistReorderInput:
      type: object
      properties:
        order:
          type: array
          items:
            type: integer
          description: Ordered list of playlist item IDs
      required: [order]
...
    SavedSearch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        filters:
          type: string
        createdAt:
          type: string
      required: [id, name, filters, createdAt]
    SavedSearchInput:
      type: object
      properties:
        name:
          type: string
        filters:
          type: string
      required: [name, filters]
...