# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaVault API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistInput {
  name: string;
  description?: string;
  isSmart?: boolean;
  smartFilters?: string;
  mediaIds?: number[];
}